updating Bao branch

+++ src/Customer.java
@@ -3,7 +3,7 @@ import java.io.IOException;
 public class Customer extends Account {
     public Customer(String cID, String name, String email, String address,
                     String phone, String customerType, String userName,
-                    String password, double totalSpending) throws IOException {
+                    String password, Long totalSpending) throws IOException {
         super(cID, name, email, address, phone, customerType, userName, password, totalSpending);
     }
